Created TODO Added more PY files for potential classes. Moved code to Monster.py from character_dnd.py
