+++ TODO.md
@@ -1,0 +1,24 @@
+# MAIN TASKs
+## 1. Create Character Class
+    Test Character Generator
+## 2. Create Monster Class
+    Create separate Python file (done).
+## 3. Create NPC Class
+    Create Separate Python file for NPC Class (done).
+## 4. Create Locations
+    Create separate file for locations (done)
+    Determine how to represent locations (dictionary? class?)
+## 5. Create Fight code
+    Partially done
+## 6. Create NPC interface code.
+## 7. Create Travel Code
+## 8. Create simple RPG
+    Character generation
+    Realm travel
+    NPC Interaction
+    Monster Fight
+
+# SUB TASKS
+
+## 1. Create Character Generator
+### Test character Generator
