+++ character_dnd.py
@@ -1,5 +1,4 @@
 import math
-from random import randint
 # Test
 
 # Class to represent a base D&D character.
@@ -250,45 +249,6 @@ class Character_DnD:
                 monster.lose_hit_points(attack_damage)
             else:
                 print(f"\n{monster.name} has successfully evaded the attack.")
-
-    def lose_hit_points(self, amount):
-        self.hit_points -= amount
-        if self.hit_points <= 0:
-            self.hit_points = 0
-            print(f"\n{self.name} has died.")
-        else:
-            print(f"\n{self.name} takes {amount} damage and has {self.hit_points} hit points remaining.")
-
-
-# Class represents a monter in DnD.
-class Monster_DnD:
-    def __init__(self, name, hit_points, armor_class, weapon, weapon_value, challenge_rating, constitution_save,
-                 attack_bonus):
-        self.name = name
-        self.hit_points = hit_points
-        self.armor_class = armor_class
-        self.weapon = weapon
-        self.weapon_value = weapon_value
-        self.challenge_rating = challenge_rating
-        self.is_dead = False
-        self.constitution_save = constitution_save
-        self.attack_bonus = attack_bonus
-
-    def __repr__(self):
-        return (f"\n {self.name} is a challenge rating {self.challenge_rating} monster with {self.hit_points} hit "
-                f"points, {self.armor_class} armor class and that wields {self.weapon}. ")
-
-    def monster_attack(self, player):
-        attack_roll = (randint(1, 20))
-        print(f"\n{self.name.title()} attempts to attack {player.name.title()} with {self.weapon}.")
-        print(f"\n{self.name.title()} rolls an attack roll of {self.attack_bonus + attack_roll} compared to an AC of "
-              f"{player.armor_class}. ")
-        if player.armor_class < self.attack_bonus + attack_roll:
-            attack_damage = self.weapon_value
-            print(f"\n{self.name.title()} has landed a hit for {attack_damage} points of damage.")
-            player.lose_hit_points(attack_damage)
-        else:
-            print(f"\n{player.name} has successfully evaded the attack.")
 
     def lose_hit_points(self, amount):
         self.hit_points -= amount
